_StylizedAOVAdaptor : Use unique driver filename
Even though the `null` driver doesn't write to a file, this is still important. It's what our renderer backend uses to coalesce multiple outputs into a single driver. When a name clashed, an output destined for the Catalogue could get coalesced into the `null` driver and never arrive in the Catalogue. There's a todo in `Globals.cpp` about detecting mismatched drivers and warning for this case - that would make this problem easier to diagnose, but not change the need for a unique name.