_StylizedAOVAdaptor : Skip pre-existing AOVs
My understanding from the Riley API is that it is OK to have the same AOV directed towards multiple display drivers. In fact, testing shows that it does work - for instance I can write the diffuse LPE to two different layers in the Catalogue at once. And various bits of RenderMan documentation imply that it was always OK in RenderMan.

But it doesn't seem to work in at least one case : the `color NPRhatchOut` AOV written from the PxrStylizedHatching display filter. If we write this to a `null` driver in addition to the Catalogue's driver, then it comes out black in the Catalogue. I wondered if there was an optimisation whereby PxrStylizedHatching skipped the write if it detected the null driver, but testing shows that not to be true - the same problem occurs if the _StylizedAOVAdaptor makes EXR drivers. And in this case the EXR ends up with valid data while the Catalogue remains black. So perhaps there is a bug whereby AOVs written by display filters can't be directed to multiple drivers?

Whatever the reason might be, the solution is fairly simple : don't add AOVs when they already exist for another driver. This makes it possible to render `color NPRhatchOut` to the Catalogue happily.