RendererAlgo : Don't mutate ObjectSets after publishing to Renderer
There is no way of signalling such a change to the Renderer without calling `ObjectInterface::link()` again, and in the meantime we were exposing the renderer to a completely inconsistent state (if it had stored the ObjectSetPtr).

We now make a copy of a set before modifying it, but will never incur more than a single copy per ObjectSet in any update cycle, because the copy is not published to the Renderer via `ObjectInterface::link()` until all calls to `LightLinks::add/removeFilterLink()` have been made.

This new immutability of ObjectSets will be used in the RenderMan backend, where we can now use ObjectSetPtr as a reliable handle for an unchanging set of objects, which makes some important optimisations possible.