CropWindowToolUI._StatusWidget : Remove broken and unused visibility control
A widget's top level visibility is controlled by the parent layout, it is not valid to control it manually.
This would result in bugs like the visibility being overwritten whenever a plug changes on the parent, and also being unable to turn visibility back on once it's off because __update is a LazyMethod with deferUntilVisible.

The correct solution to this is to make the top level widget an invisible container, and control the visibility of an internal frame that we can own fully ( ImageSelectionToolUI._StatusWidget ) ... however in this case, this visibility control is never used, so it's simpler to simply remove the broken line of code.