SpreadsheetUI : Reimplement _ClipboardAlgo
This refactors SpreadsheetUI._ClipboardAlgo to make use of Gaffer.ObjectMatrix and rely on common functionality in PlugAlgo rather than the previous use of an ObjectVector of ObjectVectors and bespoke ValueAdaptor classes for setting and converting values.

This ends up being a significant rewrite, but we preserve the existing tests where appropriate to ensure we haven't overly changed the existing behaviour.

I've mostly maintained the semantics of the previous interface used by SpreadsheetUI._PlugTableView, though the new `copyPlugs()` function copies directly to the clipboard rather than returning an ObjectMatrix to be added to the clipboard by someone else. `pasteCells()` still expects to receive an ObjectMatrix to paste, rather than pasting the contents of the clipboard directly but it serves double-duty to "paste" dropped data in `_PlugTableView.__dropFunction()` where the clipboard isn't used. The lack of symmetry here bugs me a little, but this API isn't public...