ClipboardAlgo : Add support for PathListingWidgets
This extends ClipboardAlgo to support copy/paste from/to PathListingWidgets.

Pasting is only supported to InspectorColumns as we make use of `Inspector::edit()` to set the pasted values. Copying is supported from any column with a value, though at present we only provide context menu entries and shortcuts to InspectorColumns.