CrypomatteTest : Don't put invalid manifest in metadata
The cryptomatte spec defines that manifest values should be 8 character strings, containing hexadecimal values, not floats. We are not parsing these correctly: 6.0705627102400005616e-17 is currently parsed as 6.

This was not causing test errors, because this test does not have anything to do with manifest - it tests that hashing the string put in matteNames produces the expected hash. By not passing in a manifest, we make it clearer what is being tested.