Cryptomatte : Prefer from_chars to sscanf
Now that we're using C++17, we can use from_chars, which is explicitly designed for efficient JSON parsing.

sscanf is expected to have poor performance and error handling - or at least it may. It depends on the locale, which has been known to have unexpected results for both correctness and performance.