ShaderAssignment : Move GlobalScope to ShaderPlug
This means we instantiate the GlobalScope only after using
`contextSensitiveSource()` to traverse through switches to find the
input shader. That allows a user to use `${scene:path}` to control the switch,
assigning different shaders to different locations.

This isn't completely side-effect-free, as the `${scene:path}` deletion now
applies to all ShaderPlug clients, not just ShaderAssignment. Examples of edge-case
behaviour changes include :

- Light parameters can no longer contain `${scene:path}` to receive the name of
  location created by the Light node.
- The input to an OSLImage node can no longer use `${scene:path}`, should the OSLImage
  be used by a GafferScene node.

I haven't thought of any good reason to want to do either of those things, so would
definitely be happy to accept the changes in return for the benefit of switchable
assignments. But it seems open for debate whether the likelihood of unexpected
consequences warrants this going in the next major rather than minor version. Opinions
welcome.