PlugPopup : Select text in MultiLineTextWidget
When we first made PlugPopup, MultiLineTextWidget didn't have a `setSelection()` method, but it does now. Using it removes an editing step when using the popup editors in RenderPassEditor, AttributeEditor and Spreadsheet etc. This matches the existing behaviour we had for single-line text fields.

The original code was a little unclear, but I have verified that in all cases `__firstTextWidget` either returns a TextWidget or MultiLineTextWidget, and have clarified that with a comment on the `else` block.