Cycles OIIOOutputDriver : Support `layerName` parameter
The OIIOOutputDriver and IEDisplayOutputDriver could do with a good old refactor, but I've deliberately limited myself to the most minimal change needed to get `layerName` working for OIIOutputDriver only. This does leave some discrepancies between the two drivers when a `layerName` is not specified - the OIIO driver falls back to writing "RGBA" channels whereas the IEDisplay driver uses the data name as a layer name. Changing the OIIO driver to match breaks a bunch of unit tests, so I've left that alone for now. A sensible time to approach it might be when adding multi-part EXR output support, at which point all channels would need unique names.