RenderManifest : Fix assertion failure
This occurred for debug builds in `RenderManifestTest.testReadCryptomatte` :

```
/home/john/dev/build/gaffer-1.5/include/boost/smart_ptr/intrusive_ptr.hpp:199: T* boost::intrusive_ptr<T>::operator->() const [with T = const IECore::TypedData<std::__cxx11::basic_string<char> >]: Assertion `px != 0' failed.
Aborted (core dumped)
```

It had gone unnoticed on CI, because RenderManifestTest was not being imported  by GafferSceneTest. The bug was in CacheKeyHasher, which dereferenced `key.loadedFromCryptomatteJSONStringData` without checking it was non-null.

The underlying problem that made this possible was that CacheKey had a bunch of invalid states, since it was intended to either hold a JSON string _or_ a filename and timestamp, but not both. Using individual caches for each key type allows us to remove a bunch of code and fix the bug at the same time.