RenderController : Remove ID/Path functions
These were only being called by unit tests, and I think it's highly unlikely they are used in the wild because the InteractiveRender node does not expose the RenderController. If there are any uses, it's trivial to switch to using the `renderManifest()` method anyway.

Also added missing Changelog entry for new `renderManifest()` method.