Bindings : Add missing `vector` and `optional` conversions
We were relying on `vector` and `optional` bindings being provided by the `pxr` module, but since USD moved to using its own fork of Boost::Python, these are no longer available to us.

I did spend a bit of time wondering if we should add our own `optional` and `vector` bindings. It would certainly be handy, and make a few things easier. But it's hard to know where they belong. They should really go in the lowest-level module possible, which for us would be IECorePython. But what if other projects want `vector` bindings? Then we'd have two competing bindings - their and ours - which would lead to Boost::Python warnings. Logically the only place such shared bindings could go is in Boost::Python itself. Maybe we shouldn't worry about that, because most folks are moving away from Boost anyway (but unfortunately not all to the same thing). But logically if we're going to do the bindings ourselves, we should update _all_ our bindings, and that's not really the task at hand (VFXPlatform updates are). In the short term, I think a small number of additional manual bindings as in this commit is the most pragmatic approach.