ApplicationRootTest : Fix error in Python 3.11
The error was this :

```
ERROR: testApp (GafferTest.ApplicationRootTest.ApplicationRootTest.testApp)
----------------------------------------------------------------------
DeprecationWarning: It is deprecated to return a value that is not None from a test case (<class 'GafferTest.ApplicationRootTest.ApplicationRootTest.testApp'>)
```

Python was assuming the `testApp` class was a test method because it matched the naming pattern for tests. Renaming the class to uppercase stops it being treated this way.