OSLExpressionEngine : Update for OSL 1.14
In OSL 1.13, strings were stored internally as `ustring`, and in OSL 1.14 they are now stored as `ustringhash`, which is apparently more conducive to GPU handling. But in both versions, the TypeDesc claims them to be `TypeString`. When `convert_value()` sees the source and destination types match - as they do when we convert StringData to OSL - it just clobbers the destination with bytes from the source. That works when the destination is `ustring` because it has the same layout as a `char *`, but it doesn't work when the destination is `ustringhash`, which has a different layout. So we now always do the conversion ourselves instead in this case.