ShadingEngine : Update for OSL 1.14
This one is even more fun than what we dealt with in OSLExpressionEngine in the prior commit. While OSL has moved to using `ustringhash` for strings everwhere, it _hasn't_ done so for the batched backend. So we have to do different conversions depending on whether or not we're using batched shading.