Outputs config : Add `updateInteractively` to interactive id pass
This matches how the id pass is set up in GafferSceneUI::OutputBuffer.

It is particularly necessary now that we're using bitcast ints as the ids - it seems that if the floating point values are near zero, whatever Arnold is doing to prioritize updates of non-interactive AOVs will simply ignore them, since it doesn't think they're significant.