IECoreCycles : Refactor camera handling
- Remove CameraCache. Cameras are not expensive enough or common enough to need caching.
- Never create `ccl::Camera` objects. Cycles doesn't seem to like us replacing the `Scene::camera` by pointing it at a different camera. So there is no point in creating our own `ccl::Camera` objects.
- Fix copying of our camera onto the scene camera. This was doing funky things copying back and forth with `m_cameraDefault`, and copying `shutter_table_offset` and `need_flags_update` from the previous camera. I have no idea why. Now we just copy from the relevant Cortex camera to the `Scene::camera` and call it good.
- Don't repeat ourselves. Just because we have regular cameras and dicing cameras and default cameras for each, doesn't mean we need four copies of the camera copying code. Just parameterise `updateCamera()` on the destination and call it twice.

This removes a fair bit of complexity but my initial motivation was to fix a bug : after copying camera changes to the scene camera, `Scene::camera::is_modified()` was returning `false`. It now returns `true` as you might expect. We were getting away with this bug because of _another_ bug elsewhere which caused Cycles to do unnecessary updates. But I intend to fix that.

For now I have gone out of my way to emulate bugs in the transform and projection for the default camera. They will be better off fixed separately, because they will require a bunch of adjustments to existing test cases.

I also suspect there are existing bugs in the motion transform code for cameras, and highly doubt that the shutter handling in CameraAlgo is right.
But I have kept the code intact on the off chance that it works.