Cycles Renderer : Use finer-grained locking
Rather than have one big fat mutex spanning the majority of `CyclesRenderer::render()`, lock only as necessary in the specific functions it calls. This will facilitate a small bit of mutex juggling needed in an upcoming commit.

I'm in two minds about this as it is a bit fiddly, and there are alternate approaches :

- In the upcoming commit, we could instead add an argument to `ShaderCache::get()` to indicate that the scene lock is already held.
- We could remove the lock entirely. I see no evidence that anyone else could be operating on the scene concurrently with `render()`. `Renderer::render()` can only be called single-threaded after all scene processing has been done, and I _think_ `Renderer::pause()` should return only after Cycles itself has ceased to access the scene.

I've taken the current approach because I think it helps with clarity for the mutex to be locked as close to the using code as possible, and gets us into the habit of thinking above the mutex every single time we access the scene. In the course of the current round of refactoring I've found a whole bunch of unprotected accesses that still need fixing.