IECoreCycles : Don't defer light shader assignments
We don't want to defer any assignments anyway, because it seems like an unnecessary complication. But it is more critical for lights because of a bug introduced in d5809b609f9cdab0d5ab0b8c12637d1fc46d6a5f. That moved the deletion of lights above `updateSceneObjects()` in `CyclesRenderer::render()`, meaning it was possible to attempt to assign the shader to the light _after_ the light had been deleted. I disovered this while running the tests with ASAN enabled.

I also removed the call to `light_manager->tag_update()` because there appears to be no justification for it, and the tests pass without. It's time to stop cargo-culting code around : better to have clean code that uses the Cycles API as intended, and deal with any surprises that brings by adding new tests and documenting the fix thoroughly.

I'm not sure why the same is not occurring for geometry after abe55399d85b895613ec0f6661517b2055ffe26c, but the tests run cleanly with ASAN now. In any case, I intend to remove all deferred shader assignments in a future commit.