Cycles Renderer : Don't call `delete_nodes()` with empty set
This has several consequences :

- Cycles' object manager doesn't get tagged for an unnecessary update, and we get massive improvements in update time when moving the camera in the raytraced viewport.
- Bugs elsewhere are unmasked. We were getting away with not tagging various things for update before because the object manager update was covering for us. This is dealt with with new calls to `tag_update()` in the right places.

Confession : I don't know why the call to `Mesh::tag_shader_modified()` is necessary. Without it, the shader doesn't "stick" and we get failures in `InteractiveCyclesRenderTest.testShaderEdits`, when reassigning the original red shader. I found two other ways of "fixing" this :

1. Hacking in the sledgehammer update that `delete_nodes()` would have done, but just for this one case.
2. Setting `Shader::need_update_attribute`.

I went for the `tag_shader_modified()` approach instead because it is the most localised, and seems most specific to the problem being tackled. I believe the pre-existing call to `Shader::tag_used()` is sufficient to get the shader itself compiled and onto the render device. But Cycles seems to need a kick to get the per-face shaders indices on the render device updated to point to this new shader. Pretending that the source data for those has changed seems like the most direct way we have of triggering the update.