OSL ShadingEngine : Add support for Primitive-based pointclouds
These replace the file-based pointclouds that OSL implements by default using Partio. It would have made sense to fall back to a file-based lookup if a primitive lookup failed, but that isn't possible in batched mode because BatchedRendererServices doesn't implement `pointcloud_search()` (the actual implementation is squirreled away elsewhere in liboslexec).

We think in-memory Primitive-based pointclouds are going to be way nicer to use than disk-based ones so aren't too concerned by the loss of file-based lookups.