PlugLayout : Fix bug resolving `layout:index` metadata
When an explicitly specified index coincided with an implicit index, the first item encountered won. This meant that it was impossible to put a custom widget at index zero without explicitly reordering all plugs to come after it.

The solution here is borrowed from `RenderPassEditor.__orderedColumns`, which faced the same problem.