Path : Remove `inspectionContext()`
Replacing it with a `contextProperty()` method, which is a bit more general. The ultimate goal is to be able to derive Context from RunTimeTyped and just return contexts from the regular `property()` method, but we need to approach that very carefully from a performance perspective and now is not the time.

It's worth noting that I've also moved responsibility for managing a canceller to the client : the context is no longer added by the paths. I think this makes sense since we're returning a potentially long-lived context, and there is no lifetime guarantee at all for the canceller passed to `contextProperty()`. It's also less of a worry than it was when we first added `inspectionContext()` - there are far fewer direct clients now that most clients just call `InspectorColumn::inspect()` and let it do everything.

This opens up a possibility which will soon be useful : choosing between multiple properties to provide the context.