VisualiserTool : Init m_vertexLabelStorageCapacity
We do a check in the `VertexLabel` part of the visualiser to see if the
newly required storage is greater than the
`m_vertexLabelStorageCapacity` previously allocated, before allocating
more storage. If this is not inialized to 0, it may prevent allocation.

The shared pointers that are removed from the constructor are done so
for consistency, since not all shared pointers were being constructed.
Those we can count on being initialized to `nullptr` so I've removed
them.