VisualiserTool : Bias label vertex positions
When drawing data labels for uniform variables or face indices, we use
point positions calculated from the average of the face's points. For
non-planar faces, this can result in the point being behind a face. This
biases the point positions slightly towards the camera, in clip space,
to reduce false-negative visibility tests.