InspectorColumn : Support inspectors provided by path properties
This allows a different inspector to be used on each row, giving us a different way of building views onto the scene. This will form the basis of an upcoming SceneInspector revamp to use the same Inspector framework we use everywhere else.

This requires a bunch of adjustments in `_InspectorColumn.py`, moving inspector-based tests into inner loops where we deal with cells. The one place where I've taken a bit of a liberty is where we were doing a top-level check for a SetMembershipInspector. Instead of doing a per-cell check I've instead switched to a top-level check for a _LightEditorSetMembershipColumn which I believe gives equivalent behaviour to the old code. I don't quite understand the logic here anyway - I _think_ the old code is really just a heuristic for detecting inspectors where `canEdit()` end `edit()` work with bools, but `source()` and `acquireEdit()` will yield non-boolean plugs. I expect that could be improved with a more generic per-cell check, but it seems a little involved, and outside the scope of what I'm trying to do right now. I also don't intend to use SetMembershipInspector in the SceneInspector anyway, so I think we can live with the way things are for now.