BasicInspector : Add class to simplify creation of minimal inspectors
This will be used in an upcoming SceneInspector rewrite, where we'll need a lot of different inspector types, but won't want to be writing fully-featured inspectors for each one, at least initially.

At this point, `Inspector::historyPath()` doesn't do anything useful because it skips everything where `Inspector::source()` is null. That will be addressed in a future commit.