Remove GafferCortex
We [stopped building this by default](https://github.com/GafferHQ/gaffer/commit/bbfff3f1e8bd51889cd6fff4b2c47baa11507b81) almost 6 years ago, for Gaffer 0.54.0.0. Since then it hasn't been included in any official builds. It is only used (we presume) at Image Engine, so maintenance and any future development will now be done in an internal repository there.

A couple of imports remain in `startup/browser/previews.py` and `python/GafferUI/BrowserEditor.py`, to allow Gaffer to be configured with GafferCortex functionality when it is available. It would be good to remove these too at some point, but they won't cause problems for now because they are safely protected by `try/catch`. The BrowserEditor usage in particular is a little difficult to remove, as it is used for functionality in a base class, so it seems reasonable to give us all a bit more time to get it disentangled.