Contrib : Remove `cmake`
This was a noble effort to get Gaffer building with CMake as an alternative to SCons. But it hasn't seen any updates since 2020, so is significantly out of sync with the current state of the project. We can resurrect it if someone wants to champion its development again, but for now it seems best to gracefully retire it.