Dispatchers : Remove checks for `ui:*` context variables
I've put this in a separate commit from the previous one because the consequences of a failed check are a little different for the dispatchers. For the previous commit the only likely consequence would be an unnecessary UI update, but here we'd be transmitting the bad variables to the task execution itself. Happy to omit it for now and replace with a todo if folks think that would be better.