GraphComponentBinding : Add support for aliasing children
This introduces `compatibility:childAlias:<aliasName>` metadata to provide a method for creating aliases to children of a GraphComponent by allowing `getChild( name )` to fall back to returning a child matching the value of `compatibility:childAlias:name` metadata registered to the GraphComponent when `name` isn't the name of a child of the GraphComponent. `__getItem__` and `__contains__` are also updated to use the internal `getChild()` so they also have access to aliases.

The common use-case for this is in the creation of compatibility configs when plugs are renamed between Gaffer major versions.