PathListingWidget : Fix thread safety issue
We were taking copies of the Path and PathMatcher for use in `updateWalk()`, so that the originals could be modified in the foreground while being used by the background task. But the copies themselves were being taken in the background, so there was still a small window in which data races could occur. By moving the copying operation to the UI thread, this is avoided.