HistoryWindow : Use `InspectorColumn.historyPath()`
This removes any special knowledge of specific path types, whereas before we were assuming the path had a `setContext()` method and also that it was appropriate to call that using the context from the ContextTracker. We now just track changes to the path the original InspectorColumn was showing, and get a new `historyPath()` from that each time we need one. Therefore we update every time the underlying Editor updates its path, no matter whether that edit is a context change or something else.

This does require us to add `deferUntilVisible = False` to the lazy update methods on LightEditor, RenderPassEditor and AttributeEditor, so that the history window gets updates even when the original editor is not visible. But since their PathListingWidgets won't actually perform an update until _they_ become visible, this isn't a performance concern.

This all makes it possible to use the HistoryWindow with a new Path subclass that an upcoming revamp of the SceneInspector uses.