PathListingWidget : Fix `selectionChangedSignal()` emission from `setColumns()`
We were emitting it when the new selection was visible to observers but the new columns were not - the two lists might not even have been the same length.

Also don't scroll to first selected item when updating selection - that's not appropriate when changing columns. And move `m_columnChangedConnections.clear()` just prior to where we refill it, for clarity.