PathListingWidget : Fix scrolling caused by `setColumns()`
Because we were using the nuclear option of `begin/endResetModel()`, the QTreeView would reset the vertical scrollbar to the top, because it assumed the number of rows would be changing. By taking a bit more care and declaring that we are only adding/removing columns, we get to preserve scroll state.

This also avoids flicker in the first two columns when switching tabs in the LightEditor, because those columns aren't changing at all. But it has a terrible downside - the new columns initially show the same data that the old column at that index was showing! This is no good at all. It will be fixed in the next commit. I'm just keeping that separate to make the changes more digestible by a reader.