RenderPassPlugValueWidget : Support usage outside `ScriptNode.variables`
This widget was originally designed for use with the plug returned from `ScriptNodeAlgo.acquireRenderPassPlug()`, which is the plug which specifies the current render pass for `ScriptNode::context()`. Because we had designed for this scenario only, we were getting the scene to query from the focus node. But the more common approach is to use metadata to choose an input scene on the parent node - ScenePathPlugValueWidget is a good example of that. We now support this metadata-based usage, falling back to using the focus node if the metadata is not specified.

Additionally we use a standard menu checkbox when editing regular plug, rather than the "context yellow" icon which makes sense only in the original scenario.

At the same time, we're careful to use `self.scriptNode()` rather than `plug.node().scriptNode()`, allowing the widget to be used on Editor.Settings nodes. This will be used to choose a render pass for comparison in the upcoming SceneInspector revamp.