RenderPassEditorPlugValueWidget : Remove custom size policy
This was causing the widget to get crushed to zero size in X if the parent widget wasn't providing more space explicitly. Removing it makes it act like other menu buttons - requesting a reasonable amount of space and expanding if there is additional space to be had.

I don't actually know why the custom size policy was there. It isn't necessary to allow our `layout:customWidget:renderPassSelector:width` metadata to take hold in the main menu bar (see `startup/gui/renderPassEditor.py`). And it isn't needed for the Settings window, which is the only other place where we've used the widget to date. I assume it is just a remnant from the usual battles with Qt when trying to get a specific layout behaviour.