SceneInspector : Rewrite using InspectorPath/InspectorDiffColumn
This brings the SceneInspector into line with the AttributeEditor, LightEditor and RenderPassEditor, with everything using the same basic framework. This makes EditScope-based editing and proper history tracebacks to the SceneInspector.

I wouldn't recommend reading the diff for SceneInspector.py, since this is a complete rewrite. Better just to read the code as if it were completely new.