Switch : Use internal connection when switch is disabled
Previously enabled this optimisation only when we could determine that both the `enabled` and `index` plugs had static (non-computed) values. Here we take it one step further : if we can determine that `enabled` has a static value of `false` then we can use the same optimisation even when `index` might be non-static, because the index will never be evaluated.