Arnold ShaderNetworkAlgo : Improve naming of Arnold shader nodes
Previously, the output shader (the one most likely to be useful to people) was named only with a stringified hash, which was next to useless for identification purposes. We now precede the hash with the corresponding shader handle for all generated shader nodes. Since the handle comes from the node name (when assigned in Gaffer) or prim name (when loaded from USD), this makes it easier for a user to identify the shader.

The rules around the `name` parameter to `ShaderNetworkAlgo::convert()` are a little convoluted so that we can continue to name light and light filters nodes exactly as before, without including a shader handle. The alternative would be to have two parameters, one `nameToGiveToTheOutputNode` and another `nameFormatToUseForTheOtherNodes`. This felt even more convoluted, so I went for the simpler approach, which also happens to maintain backwards compatibility for existing users of the API.