Arnold plugin : Fix conflict with Arnold's internal allocator
`libai.so` exports symbols for `operator new` and `operator delete`, implemented using its own internal memory allocator. Our `Gaffer.so` plugin with the output driver in it was binding to these symbols in preference to the standard ones. This lead to crashes caused by objects getting allocated with one allocator (an object created in IECore for example) and deleted by another (when the plugin disposed of the object).

This problem only came to light when we updated from GafferHQ/dependencies `10.0.0a1` to `10.0.0a2`, with the relevant change being an upgrade to OpenImageIO 3. This contained the following change, which meant we were no longer loading the `osl.imageio` plugin :

https://github.com/AcademySoftwareFoundation/OpenImageIO/pull/4245

Why would that be relevant? Because OpenImageIO loads all its plugins with `RTLD_GLOBAL` and was therefore was putting `libstdc++` in the global symbol table, ahead of `libai` in the search order. With the plugin no longer loading, that was no longer happening and the problem was revealed.

This could be fixed by using `OPENIMAGEIO_PLUGIN_PATH` to get `osl.imageio` loading again. But we don't use the plugin anyway, and it pollutes the global symbol table with a huuuuge amount of stuff in addition to `libstdc++`, and we'd be better off without that. So instead we fix by linking our plugin to `libstdc++` explicitly, making sure it comes before `libai.so` in the link order.

See also https://github.com/GafferHQ/dependencies/pull/284 where I remove the `osl.imageio` from our dependencies entirely.