Menu : Flag existing submenus for deletion when building menu
We were calling `qtMenu.clear()` to clear the menu on each rebuild, but this does not cause submenus and their actions to be cleaned up, leading to an accumulation of QActions every time a menu is opened. This accumulation increased the expense of every new QAction created, which wasn't too noticeable until our upgrade to Qt 6 where this became very apparent in Node menu update lag, especially when searching. We now expressly flag to Qt that the old QMenus can be deleted when we clear the menu and are about to build their replacements, this reduces much of the lag introduced by the accumulation of QActions across multiple openings of the node menu, though doesn't address the potential lag from QAction accumulation within a single search, but this will be addressed in the next commit.