Menu : Cache search menu actions
We were creating new QActions for each update of the search menu, this would result in potentially thousands of actions being created during a search, many of which were duplicates. This caused the search menu to feel increasingly laggy for longer searches where more actions are created as the overhead for creating each new action increased, most noticeably in Qt 6. We now cache and reuse the actions across a single search, which vastly improves interactivity for longer searches.