PathListingWidget : Reset stretch column proportions on column change
Usually when we resize `Stretch` columns we preserve the ratio of column widths as the user may have chosen to adjust them before resizing the PathListingWidget. Though when we change the columns in an existing PathListingWidget, those existing widths are less relevant as the number of, order of, and columns themselves could be entirely different so we now ignore any existing header widths and instead give each stretch column an equal share of the available space. This is a little heavy-handed but prevents situations such as where calling `setColumns()` to switch from one stretch column to two would result in the first new stretch column being given the majority of the available space.