ContextQueryUI : Don't create queries for unsupported plug types
Better to never make the node than to make a node which doesn't work. In the case of NameValuePlugs, the created node sortof almost kindof worked, but in a confusing way where dependency tracking was broken. That could maybe be fixed in the node itself, but there's not really any point because there's no associated data type that could be used for such a context variable anyway.