OpenImageIOReader : Disable channelName heuristics for single-part files
The `Default` mode tries to reverse engineer whatever weird logic Nuke uses for
writing and reading non-standards-compliant exrs. A part of Nuke's logic that
we previously weren't incorporating is that it ignores part names set on
single-part files. ( In standards compliant exrs, the part names never do
anything to channel names, so this isn't an issue when channelInterpretation
is set to "EXR Specification" )