TestShader : Improve output handling
It made no sense to force a colour output for every type of shader, and having it meant we couldn't test outputs of other types, or shaders with multiple outputs. Outputs are now always children of the `out` plug, as we intend to be standard for all Shader subclasses (although we are not there with ArnoldShader yet - see #5542).

I haven't worried about adding a shim for backwards compatibility since this node is intended for testing only.